Add shortcut to run some tests, with pipfile for easy setup
- make test to run tests against a local clickhouse container
- add pipfile to install virtual env easily
- add requests as project dependency in setup.py and Pipfile
- make simpletests.py discoverable by unittest
