+++ Pipfile
@@ -1,0 +1,14 @@
+[[source]]
+name = "pypi"
+url = "https://pypi.org/simple"
+verify_ssl = true
+
+[dev-packages]
+autopep8="*"
+pylint="*"
+
+[packages]
+requests="*"
+
+[requires]
+python_version = "2.7"
